% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_functions.R
\name{check_inputs}
\alias{check_inputs}
\title{Helper function to check inputs}
\usage{
check_inputs(input_list, x)
}
\arguments{
\item{input_list}{list of inputs variables}

\item{x}{input dataframe}
}
\description{
Function for checking varnames passed as strings
}
\keyword{internal}
