% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_functions.R
\name{generate_slp_ensemble}
\alias{generate_slp_ensemble}
\title{Generate simulated linear pooling ensemble}
\usage{
generate_slp_ensemble(
  slp_predictions,
  qp_var = "predictions",
  byvars = NULL,
  quantiles = c(0.01, 0.025, seq(0.05, 0.95, 0.05), 0.975, 0.99)
)
}
\arguments{
\item{slp_predictions}{This is the set of simulated predictions returned by \code{generate_slp_predictions}; it
must contain a column named "predictions"}

\item{qp_var}{string name of the resulting combined variable (defaults to "predictions")}

\item{byvars}{(default = NULL), a character vector indicating the variables that uniquely define
a set of predictions}

\item{quantiles}{a numeric vectors of values between 0 and 1 exclusive; these are the quantile levels to
estimate. The default set contains 23 levels (\code{c(0.01,0.025,seq(0.05,0.95,0.05),0.975,0.99)})}
}
\value{
Returns a data table of the simulated linear pooling ensemble
}
\description{
This function takes a set of simulated predictions (for example, returned by \code{generate_slp_predictions},
and returns the estimates of the quantile function Q(p) for the set. Optionally, multiple sets of
simulated predictions can be given, with sets distinguished by \code{byvars} parameter
}
\examples{
generate_slp_ensemble(slp_predictions = slp)
generate_slp_ensemble(slp_predictions, byvars=c("location","date"),quantiles=c(0.025,0.25,0.5,0.75,0.975))
}
