% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_functions.R
\name{generate_vinc_ensemble}
\alias{generate_vinc_ensemble}
\title{Generate vincentization / quantile-over-quantile ensemble}
\usage{
generate_vinc_ensemble(
  x,
  qp_var,
  p_var,
  byvars = NULL,
  pooltype = c("mean", "median")
)
}
\arguments{
\item{x}{This is the data frame/data.table that contains the \code{p_var} and \code{qp_var} and any
grouping variables \code{byvars} (i.e. defining set of quantile functions).}

\item{qp_var}{This is the name of the variable that holds the estimate of the number
of hospitalizations.}

\item{p_var}{This is the name of the variable that holds the probability, p, \code{qp_var} will be pooled
over equal levels of \code{p_var}}

\item{byvars}{(default = NULL), a character vector indicating which variables distinguish different
ensembles}

\item{pooltype}{a string indicating whether "mean" (default) or "median" of the quantile levels should
be returned}
}
\value{
Returns a data table of the quantile-by-quantile or vincentization ensemble
}
\description{
This function takes a set of quantile functions, and returns the mean (default) or median
quantile level over the quantiles. This is also known as the vincentization approach.
}
\examples{
generate_vinc_ensemble(hosp, "value")
generate_vinc_ensemble(hosp, target_variable="val", pooltype="median")
generate_vinc_ensemble(hosp, target_variable="val", pooltype="median", byvars=c("loc","date","quantile")
}
