% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_functions.R
\name{get_parallel_status}
\alias{get_parallel_status}
\title{Helper function to set up parallelization config}
\usage{
get_parallel_status(nmods, parallel, clusters)
}
\arguments{
\item{nmods}{integer number of models}

\item{parallel}{string, one of "auto","on","off"}

\item{clusters}{integer number of threads (or NULL)}
}
\description{
Function sets up parallelization configuration list
}
\keyword{internal}
