% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_functions.R
\name{pre_smooth_quantile_functions}
\alias{pre_smooth_quantile_functions}
\title{Pre-smooth individual model quantile function, prior to applying the SLP method}
\usage{
pre_smooth_quantile_functions(
  x,
  qp_var,
  p_var,
  time_var,
  byvars = NULL,
  smooth_col_name = paste0(qp_var, "_hat"),
  loess_span = 0.75
)
}
\arguments{
\item{x}{the source data table containing the quantile function(s) over some unit, generally time}

\item{qp_var}{The name of the variable holding the $Q(p)$ values}

\item{p_var}{The name of the variable in \code{x} holding the p values}

\item{time_var}{The name of the variable in \code{x} holding the unit (i.e. time) over which you want to pre-smooth Q(p) for
fixed values of \code{p_var}}

\item{byvars}{The vector of additional (default =NULL) variables in \code{x} that define additional groups}

\item{smooth_col_name}{string name of new column that will contain the smoothed Q(p) values; the default is to
produce a column name \verb{<qp_var>_hat} (unless it already exists}

\item{loess_span}{numeric value between 0 and 1 (default=0.75) indicating the span of data which will be used in the
locally weighted regression process}
}
\value{
Returns the a data table wih \code{qp_var},\code{p_var},\code{time_var},\code{byvars} (if any), and the new \code{smooth_col_name}
}
\description{
This function takes in a set of individual models, and pre-smooths the quantile
function to even out day-to-day volatility, prior to inserting into the SLP
prediction process. The function using a locally weight regression smoothing approach
}
\examples{
pre_smooth_quantile_functions(x=hosp,qp_var="hospitalizations",p_var="quantile", time_var="target_end_date", byvars=c("src","location"), smooth_col_name="value_sm")
}
