% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_functions.R
\name{trim_vector}
\alias{trim_vector}
\title{Trim a vector by a given level}
\usage{
trim_vector(x, trim_level = 0.05, tag = F)
}
\arguments{
\item{x}{the vector of values}

\item{trim_level}{numeric between 0 and 1, default=0.05}

\item{tag}{logical default F, set to T to return a logical vector rather than
the trimmed version of x}
}
\description{
Function takes a vector \code{x} and either returns a logical vector of length
equal to length(x), indicating indices of x that would be kept given
a trim level (see below), or returns a trimmed version of x itself. The
trim level is provided between 0 and 1, and trim_level/2 and 1-trim_level/2
indicate the quantile values at which x will be trimmed (or tagged)
}
\examples{
trim_vector(x,trim_level=0.1)
trim_vector(my_values, tag=T)
}
